Replace MD_EXTENSIONS list with MARKDOWN dict in pelicanconf
Apparently this moved at some point? Anyhow, enabled the `abbr`
extension for adding `<abbr title="...">` around acronyms; `tables` for
table support with the syntax used on GitHub et al.; and `toc` with the
`anchorlink` option turned on so that H1-H6 tags automatically get an
anchor for linking to. Will consider turning on `codehilite` at some
point, but I need to add CSS rules for it if I do.